TP2-React-Terminado

+++ TP2-React/TP2-Final/src/paginas/empleados/BajaEmpleado.jsx
@@ -1,44 +1,57 @@
-import React, { useState } from 'react'
 import BarraNavegacion from '../../componentes/BarraNavegacion';
 import JSON_SERVER_URL from '../../json-server-config';
 import '../../styles/EmpleadoStyle.css'
+import nullInputValidation from '../../utils/validations';
+import useForm from '../../hooks/useForm';
+import { useState } from 'react';
 
 const BajaEmpleado = () => {
-
-    const [inpudData, setInputData]=useState("");
-    const url_delete = JSON_SERVER_URL+'/'+inpudData;
   
-    const buttonClick = () =>{
+    const initialValues = { id: '' };
+    const { values, errors, handleChange, handleSubmit } = useForm(initialValues, nullInputValidation);
+    const [mensaje, setMensaje]=useState("");
 
-         const requestOptions = {
-          method: 'DELETE',
-          headers: { 'Content-Type': 'application/json' },
-        };
-        fetch(url_delete, requestOptions)
-
-            .then(response => {
-              if (response.ok) {
-                console.log('Recurso eliminado con éxito');
-                return response.json(); 
-              } else {
-                console.error('Error al eliminar el recurso');
-              }
-            })
-            .catch(error => console.error('Error:', error));
+    const mensajeFunction=(texto)=>{
+      setMensaje(texto)
+      const intervalo = setTimeout(() => {
+        setMensaje("")
+      }, 2000);
     }
 
-    const inputHandler = (event) =>{
-      setInputData(event.target.value);        
-    }
+    const buttonClick = () => {    
+      
+      handleSubmit().then(() => {
+          if (Object.keys(errors).length === 0) {
+              const url_delete = JSON_SERVER_URL + '/' + values.id;
+
+              const requestOptions = {
+                  method: 'DELETE',
+                  headers: { 'Content-Type': 'application/json' },
+              };
+
+              fetch(url_delete, requestOptions)
+                  .then(response => {
+                      if (response.ok) {
+                          mensajeFunction("Empleado eliminado")
+                          return response.json();
+                      } else {
+                        mensajeFunction("Error al eliminar empleado")
+                      }
+                  })
+                  .catch(error => console.error('Error:', error));
+          }else{mensajeFunction("Campos vacíos")}
+      });
+  };
 
   return (
-    <div >
+    <div className='EmpleadoStyle_fondo'>
       <BarraNavegacion></BarraNavegacion>
       <div className='EmpleadoStyle_contenedor'>        
         <label className='EmpleadoStyle_label'> Eliminar por id: </label>
-        <input className='EmpleadoStyle_input' type='text' onChange={inputHandler} value={inpudData} ></input>
+        <input className='EmpleadoStyle_input' name='id' type='text' onChange={handleChange} value={values.id} ></input>
         <br></br>
         <button className='EmpleadoStyle_button' onClick={buttonClick}>Eliminar</button>
+        <p className='EmpleadoStyle_mensaje'>{mensaje}</p>
       </div>
     </div>
   )
